requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
python-escpos==3.0a9
python-barcode==0.15.1
Pillow==10.0.1
pynput==1.7.6
requests==2.31.0